+++ blockchain/client.go
@@ -124,7 +124,6 @@ func (c *Client) Health() (*tmctypes.ResultHealth, error) {
 func (c *Client) GenesisValidators() ([]*tmtypes.Validator, error) {
 	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
 	defer cancel()
-
 	return c.clt.GenesisValidators(ctx)
 }